IPv6 support and listening to custom addresses
This patch adds IPv6 support for obigstore as well as the ability for
the obigstore server to listen to custom (user-defined) IP addresses
or hostnames.

Now Unix.getaddrinfo is used to obtain sockaddrs so that users can
supply domain names and services names as well, instead of being
forced to use IP addresses and integer port numbers.

This fixes #3 and #4

Signed-off-by: Vincent Bernardoff <vb@luminar.eu.org>