Update stream must be pushed into asynchronously until it's accessed.
When a slave attaches itself to the master, the latter must not wait for
update ACK until the former has finished fetching the initial dump.
When the slave accesses the update stream, the stream is switched to Sync mode
and from now on the master will wait for ACK from the slave after each push.

Also, add a No_stream mode (updates not streamed, only interested in data
files) for use by  ob_dump -serverside  and DUMP in ob_repl.